requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
Flask-WTF==1.1.1
Flask-Talisman==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
eventlet==0.33.3